move addEventListener to root page
