+++ app/client/src/pages/AppViewer/AppPage.tsx
@@ -23,20 +23,6 @@ type AppPageProps = {
 
 export function AppPage(props: AppPageProps) {
   useDynamicAppLayout();
-  const messageHandler = (event: MessageEvent) => {
-    if (event.currentTarget !== window) return;
-    if (event.type !== "message") return;
-    if (!isValidDomain(event.origin)) return;
-    const storeKey = `APPSMITH_LOCAL_STORE-${props.pageId}`;
-    const storageKeys = JSON.parse(localStorage.getItem(storeKey) || "{}");
-    const data = JSON.parse(event.data);
-    for (const key in data) {
-      storageKeys[key] = data[key];
-    }
-    const storeString = JSON.stringify(storageKeys);
-    localStorage.setItem(storeKey, storeString);
-  };
-  window.addEventListener("message", messageHandler);
   useEffect(() => {
     // add postmessage
     AnalyticsUtil.logEvent("PAGE_LOAD", {
@@ -55,29 +41,4 @@ export function AppPage(props: AppPageProps) {
   );
 }
 
-function isValidDomain(domain: string): boolean {
-  const regex1 = new RegExp("/(.+?)[.]manabie.com$");
-  const regex2 = new RegExp("/(.+?)[.]web.app$");
-  const regex3 = new RegExp("/(.+?)[.]manabie.io$");
-  const regex4 = new RegExp("/(.+?)[.]manabie.net$");
-  if (
-    (window.location.origin == "http://localhost" ||
-      window.location.origin ==
-        "https://appsmith.local-green.manabie.io:31600" ||
-      regex3.test(window.location.origin)) &&
-    domain.indexOf("localhost") > -1
-  ) {
-    return true;
-  }
-
-  if (
-    regex1.test(domain) ||
-    regex2.test(domain) ||
-    regex3.test(domain) ||
-    regex4.test(domain)
-  ) {
-    return true;
-  }
-  return false;
-}
 export default AppPage;
